+++ src/routes/Index/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Index extends React.Component {
+  render() {
+    return <div>index page</div>
+  }
+}
+
+export default Index;