+++ src/routes/Questions/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Questions extends React.Component {
+  render() {
+    return <div>questions page</div>
+  }
+}
+
+export default Questions;