+++ src/routes/Realtime/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Realtime extends React.Component {
+  render() {
+    return <div>realtime page</div>
+  }
+}
+
+export default Realtime;