+++ src/routes/Reward/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+class Reward extends React.Component {
+  render() {
+    return <div>reward page</div>
+  }
+}
+
+export default Reward;